ci: add `localization/` linting
# Motivation

In order to ensure that we prefix localization
keys with the correct prefix `br_`, we need to
have a more automatic way of checking this.

Since the localization keys are stored in `YAML`
files, it should be pretty easy to simply check
the keys in those files with a simple script.

However, the code in many of those files is not
compliant with the `YAML` standard, so we need to
fix that first.

# Changes

This commit adds a Github Action workflow that
ensures that the `YAML` files in the
`localization/` does not have syntax errors and
follow the configuration in
`scripts/lint/.yamllint.yml`.

# Future work

A bunch of files are not compliant, and should be
removed from the `--paths-to-ignore` list.