fix: formation cannot be mobilized
This commit solves the following error:

```
[jomini_script_system.cpp:262]: Script system
error!
Error: fully_mobilize_army effect [ Scope
formation cannot be mobilized ]
Script location: file: common/history/
military_deployments/00_military_deployments.txt
line: 3
```
and similar errors for other formations.

The fix was to just delete the mobilization code,
since it was not working anyway. For most of the
initial wars in the game, none of the formations
are in the correct place anyway.