refactor: move modifiers from tech
# Motivation

As far as I know, technology is the only way of
disabling colonial modifiers such as malaria or
hostile climate. So if early colonization should
be enabled, we still need to keep technologies
such as `new_world_colonial`.

But the modifiers they provide can still be made
visible as a separate modifier.