feat: add new production tech era
# Motivation

Since the mod spans an additional 100 years
compared to the base game, the tech tree should
also be expanded to reflect this.

Creating 100 years of tech in 3 categories in one
go is unfortunately too much work. So just adding
one tech category for a single era is a good
start.

# Changes

- Adjusted the eras of techs so that era 5 becomes
  era 6, era 4 becomes era 5, etc
- Added a new era 1, which contains the very basic
  military and society techs
- Added a new era 1 production techs
- Some era 2 techs are now era 1 techs, such as
  `sericulture` and `enclosure`
- Adjusted the starting techs so that there are
  not too many changes actual changes

# Future work

- Add distinct icons for each of the new techs
- Add PMs to reflect artisans manufacturing
  better.
- Add techs for the other two categories