fix: use laws allowed by techs
This commit solves errors such as:

```
[country_law_manager.cpp:359]: Country IRE is not
permitted to retain law Wealth Voting, fix this or
consider Autocracy instead
```

This commit does not solve every starting tech
error, but it does solve quite a few.

If the starting techs after all were correct, then
the technologies should be changed instead. Having
hundreds of errors in the log is not helpful.