Revert "fix: removed unchanged moved vanilla files"
This reverts commit 8398943e34657cc123def123ca016e7122c2f690.

The reason why this commit was reverted is that
the logic is in use. For example the PMs will
naturally overflow under the subsidize button
without these files.

A follow up commit should add some description of
what this code is supposed to do, so we do not
accidentally delete it again.