Elysium starting religion
Elysium has been struggling to attract migrants in 1.8 due to their starting religion. I switched their starting religion to Elysian Orthodox and switched the decision to go Hellenic to make it possible for them at the start of game this doesn't effect BR which doesn't get the decision until after the Olympics are founded.