renamed file
renamed romaion_military_formations to ere_military_formations. I will use ERE or BR_ERE as the main internal reference for the nation of Basileia Romaion in files going forward as it easier to find and organize the tag will remain BYZ the change is just for file names and internal memos.